refactored the contactlist
