+++ styles/main.scss
@@ -43,6 +43,7 @@ hr.intro-break {
       .sub-header{
         font-size: 30px;
         padding-left: 5px;
+        font-family: monospace;
       }
     }
   }
